reorganize multi-scale feature transform functions into base_head
